Adding unit tests for autoscaling leaf
